+++ external/anomaly/adapters/__init__.py
@@ -1,8 +1,6 @@
-"""
-Configurable parameters for Padim anomaly segmentation task
-"""
+"""Adapters for Anomalib."""
 
-# Copyright (C) 2021 Intel Corporation
+# Copyright (C) 2022 Intel Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,13 +13,3 @@ Configurable parameters for Padim anomaly segmentation task
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions
 # and limitations under the License.
-
-from adapters.anomalib.configs.padim import PadimAnomalyBaseConfig
-from attr import attrs
-
-
-@attrs
-class PadimAnomalySegmentationConfig(PadimAnomalyBaseConfig):
-    """
-    Configurable parameters for PADIM anomaly segmentation task.
-    """
