+++ external/anomaly/adapters/anomalib/callbacks/__init__.py
