+++ external/anomaly/adapters/anomalib/callbacks/progress.py
