+++ external/anomaly/adapters/anomalib/callbacks/score_report.py
