+++ external/anomaly/adapters/anomalib/configs/anomalib_config.py
