+++ external/anomaly/adapters/anomalib/configs/base/__init__.py
