+++ external/anomaly/adapters/anomalib/configs/configuration_enums.py
