+++ external/anomaly/adapters/anomalib/configs/padim/__init__.py
