+++ external/anomaly/adapters/anomalib/configs/stfpm/__init__.py
