+++ external/anomaly/adapters/anomalib/configs/stfpm/configuration.py
@@ -16,9 +16,9 @@ Configurable parameters for STFPM anomaly base task
 # See the License for the specific language governing permissions
 # and limitations under the License.
 
+from adapters.anomalib.configs.base import BaseAnomalyConfig
+from adapters.anomalib.configs.configuration_enums import EarlyStoppingMetrics
 from attr import attrs
-from ote_anomalib.configs.base import BaseAnomalyConfig
-from ote_anomalib.configs.configuration_enums import EarlyStoppingMetrics
 from ote_sdk.configuration.elements import (
     ParameterGroup,
     add_parameter_group,
