+++ external/anomaly/adapters/anomalib/data/__init__.py
