+++ external/anomaly/adapters/anomalib/exportable_code/anomaly_detection.py
