+++ external/anomaly/adapters/anomalib/exportable_code/anomaly_segmentation.py
