+++ external/anomaly/tests/test_configurable_parameters.py
@@ -20,10 +20,11 @@ import logging
 import os
 
 import pytest
+from adapters.anomalib.configs import get_anomalib_config
 from configs.anomaly_classification.padim import PadimAnomalyClassificationConfig
 from configs.anomaly_classification.stfpm import STFPMAnomalyClassificationConfig
-from ote_anomalib.configs import get_anomalib_config
 from ote_sdk.configuration.helper import convert, create
+
 from tests.helpers.config import get_config_and_task_name
 
 logger = logging.getLogger(__name__)
